Move some basic 'tractanalysis' method to 'tractogram.streamline_operations'
