+++ scilpy/tractanalysis/reproducibility_measures.py
@@ -90,73 +90,6 @@ def binary_classification(segmentation_indices,
     return sensitivity, specificity, precision, accuracy, dice, kappa, youden
 
 
-def get_endpoints_density_map(streamlines, dimensions, point_to_select=1):
-    """
-    Compute an endpoints density map, supports selecting more than one points
-    at each end.
-    Parameters
-    ----------
-    streamlines: list of ndarray
-        The list of streamlines to compute endpoints density from.
-    dimensions: tuple
-        The shape of the reference volume for the streamlines.
-    point_to_select: int
-        Instead of computing the density based on the first and last points,
-        select more than one at each end. To support compressed streamlines,
-        a resampling to 0.5mm per segment is performed.
-    Returns
-    -------
-    ndarray: A ndarray where voxel values represent the density of endpoints.
-    """
-    endpoints_map_head, endpoints_map_tail = \
-        get_head_tail_density_maps(streamlines, dimensions, point_to_select)
-    return endpoints_map_head + endpoints_map_tail
-
-
-def get_head_tail_density_maps(streamlines, dimensions, point_to_select=1):
-    """
-    Compute two separate endpoints density maps for the head and tail
-    Parameters
-    ----------
-    streamlines: list of ndarray
-        The list of streamlines to compute endpoints density from.
-    dimensions: tuple
-        The shape of the reference volume for the streamlines.
-    point_to_select: int
-        Instead of computing the density based on the first and last points,
-        select more than one at each end. To support compressed streamlines,
-        a resampling to 0.5mm per segment is performed.
-    Returns
-    -------
-    A tuple containing
-        ndarray: A ndarray where voxel values represent the density of
-            head endpoints.
-        ndarray: A ndarray where voxel values represent the density of
-            tail endpoints.
-    """
-    endpoints_map_head = np.zeros(dimensions)
-    endpoints_map_tail = np.zeros(dimensions)
-    for streamline in streamlines:
-        nb_point = max(2,  int(length(streamline))*2)
-        streamline = set_number_of_points(streamline,
-                                          nb_point)
-        points_list_head = \
-            list(streamline[0:point_to_select, :])
-        points_list_tail = \
-            list(streamline[-point_to_select:, :])
-        for xyz in points_list_head:
-            x_val = np.clip(xyz[0], 0, dimensions[0]-1).astype(int)
-            y_val = np.clip(xyz[1], 0, dimensions[1]-1).astype(int)
-            z_val = np.clip(xyz[2], 0, dimensions[2]-1).astype(int)
-            endpoints_map_head[x_val, y_val, z_val] += 1
-        for xyz in points_list_tail:
-            x_val = np.clip(xyz[0], 0, dimensions[0]-1).astype(int)
-            y_val = np.clip(xyz[1], 0, dimensions[1]-1).astype(int)
-            z_val = np.clip(xyz[2], 0, dimensions[2]-1).astype(int)
-            endpoints_map_tail[x_val, y_val, z_val] += 1
-    return endpoints_map_head, endpoints_map_tail
-
-
 def approximate_surface_node(roi):
     """
     Compute the number of surface voxels (i.e. nodes connected to at least one
